+++ tests/sdk/testPixDisputeLog.py
@@ -1,0 +1,47 @@
+import starkinfra
+from unittest import TestCase, main
+from tests.utils.user import exampleProject
+
+
+starkinfra.user = exampleProject
+
+
+class TestPixDisputeLogQuery(TestCase):
+
+    def test_success(self):
+        logs = list(starkinfra.pixdispute.log.query(limit=10))
+        logs = list(starkinfra.pixdispute.log.query(
+            limit=10,
+            dispute_ids={log.dispute.id for log in logs},
+            types={log.type for log in logs}
+        ))
+        self.assertEqual(10, len(logs))
+        print("Number of logs:", len(logs))
+
+
+class TestPixDisputeLogPage(TestCase):
+
+    def test_success(self):
+        cursor = None
+        ids = []
+        for _ in range(2):
+            logs, cursor = starkinfra.pixdispute.log.page(limit=2, cursor=cursor)
+            for log in logs:
+                print(log.dispute.transactions[0])
+                self.assertFalse(log.id in ids)
+                ids.append(log.id)
+            if cursor is None:
+                break
+        self.assertTrue(len(ids) == 4)
+
+
+class TestPixDisputeLogInfoGet(TestCase):
+    def test_success(self):
+        logs = starkinfra.pixdispute.log.query()
+        log_id = next(logs).id
+        log = starkinfra.pixdispute.log.get(id=log_id)
+        self.assertEqual(log_id, log.id)
+
+
+if __name__ == '__main__':
+    main()